backend/requirements.txt
Flask==3.1.3
flask_cors==6.0.2
pydantic==2.12.5
pytest==9.0.2
python-dotenv==1.2.2
Requests==2.32.5
